FIxing tests and refining codebase (formatting, linting and minor bug fixes)

+++ src/rs/fft/mod.rs
@@ -9,12 +9,12 @@ pub use types_rust::SAMPLE_RATE;
 /// Get current timestamp in milliseconds since Unix epoch
 /// This function works on both WASM and native builds
 pub fn now_millis() -> i64 {
-    #[cfg(target_arch = "wasm32")]
-    {
-        js_sys::Date::now() as i64
-    }
-    #[cfg(not(target_arch = "wasm32"))]
-    {
-        chrono::Utc::now().timestamp_millis() as i64
-    }
+  #[cfg(target_arch = "wasm32")]
+  {
+    js_sys::Date::now() as i64
+  }
+  #[cfg(not(target_arch = "wasm32"))]
+  {
+    chrono::Utc::now().timestamp_millis()
+  }
 }
